refactor(commands/suggest): format, simplify some stuff, remove timeouts

+++ src/commands/suggest.ts
@@ -5,7 +5,6 @@ import Command from "../struct/Command"
 import Suggestion from "../entities/Suggestion"
 import Roles from "../util/roles"
 import TextChannel from "../struct/discord/TextChannel"
-import { config } from "winston"
 
 export default new Command({
     name: "suggest",
@@ -14,27 +13,11 @@ export default new Command({
     permission: Roles.ANY,
     usage: "['anon'] <title> | <body> | [team/s]",
     async run(this: Command, client: Client, message: Message, args: Args) {
-        var filteredTimeouts = client.timeouts.filter(x => x.id === message.author.id)
-        if(filteredTimeouts.length === 1) {
-            if((Date.now()) > (filteredTimeouts[0].time + 600000)) {
-                client.timeouts.splice(client.timeouts.findIndex(x => x.id === message.author.id),1)
-            } else {
-                var date = new Date()
-                date.setTime(filteredTimeouts[0].time+600000)
-                if(message.channel.type != "dm")
-                    message.delete()
-                const erMessage = await message.channel.sendError(`You can not send another suggestion until ${date.toUTCString()}.`)
-                return await erMessage.delete({ timeout: 10000 })
-            }
-        }
-        const anon = args.consumeIf(a => ["anon", "anonymous"].includes(a.toLowerCase()))
-        var staff = false
-        if(message.guild) {
-           staff = message.guild.id === client.config.guilds.staff
-        }
-        
+        const anon = args.consumeIf(["anon", "anonymous"])
+        const staff = message?.guild.id === client.config.guilds.staff
+
         const suggestionsChannel = client.config.suggestions[staff ? "staff" : "main"]
-        if (message.channel.id !== suggestionsChannel && message.channel.type != "dm")
+        if (message.channel.id !== suggestionsChannel && message.channel.type !== "dm")
             return message.channel.sendError(
                 `Please run this command in <#${suggestionsChannel}>!`
             )
@@ -48,15 +31,13 @@ export default new Command({
         if (!body) error = "You must specify a suggestion body!"
 
         if (error) {
-            if(message.channel.type !== "dm") {
-                await message.delete()
-            }
+            if (message.channel.type !== "dm") await message.delete()
             const errorMessage = await message.channel.sendError(error)
             return await errorMessage.delete({ timeout: 10000 })
         }
 
         // delete message asap if suggestion is anonymous
-        if (anon && message.channel.type != "dm") await message.delete()
+        if (anon && message.channel.type !== "dm") await message.delete()
         else message.react("👌")
 
         const suggestion = new Suggestion()
@@ -67,23 +48,18 @@ export default new Command({
         suggestion.body = body
         suggestion.teams = teams || null
         suggestion.staff = staff
-        var sugChannel
-        if(message.channel.id === client.config.suggestions.staff) {
-            sugChannel = client.guilds.cache.get(client.config.guilds.staff).channels.cache.get(suggestionsChannel)
-        } else {
-            sugChannel = client.guilds.cache.get(client.config.guilds.main).channels.cache.get(suggestionsChannel)
-        }
-        
 
-        const suggestionMessage = await (sugChannel as TextChannel).send({
-            embed: await suggestion.displayEmbed(client)
-        })
-        client.timeouts.push({id:message.author.id,time: new Date().getTime()})
+        const category = staff ? "staff" : "main"
+        const suggestionsID = client.config.suggestions[category]
+        const suggestions = <TextChannel>await client.channels.fetch(suggestionsID, true)
+
+        const embed = await suggestion.displayEmbed(client)
+        const suggestionMessage = await suggestions.send({ embed })
         suggestion.message = suggestionMessage.id
         await suggestion.save()
 
         await suggestionMessage.react(client.config.emojis.upvote)
         await suggestionMessage.react(client.config.emojis.downvote)
-        if ((!anon) && message.channel.type != "dm") await message.delete()
+        if (!anon && message.channel.type !== "dm") await message.delete()
     }
 })
